remove extra types
