+++ internal/cli/views/items/creditcard/show/list.go
@@ -11,13 +11,12 @@ import (
 
 type modelList struct {
 	list       list.Model
-	ready      bool
 	nextPage   func()
 	allFetched bool
 	Control
 }
 
-func (m modelList) getMoreItems() []list.Item {
+func (m *modelList) getMoreItems() []list.Item {
 	items := []list.Item{}
 	itemsPerPage := m.list.Paginator.PerPage
 	offset := len(m.list.Items())
@@ -36,16 +35,11 @@ func (m modelList) getMoreItems() []list.Item {
 	return items
 }
 
-func (m modelList) Init() tea.Cmd {
+func (m *modelList) Init() tea.Cmd {
 	return m.list.SetItems(m.getMoreItems())
 }
 
-func (m modelList) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	if !m.ready {
-		cmd := m.list.SetItems(m.getMoreItems())
-		m.ready = true
-		return m, cmd
-	}
+func (m *modelList) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch msg := msg.(type) {
 	case tea.WindowSizeMsg:
 		m.list.SetWidth(msg.Width)
@@ -54,7 +48,6 @@ func (m modelList) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.KeyMsg:
 		switch keypress := msg.String(); keypress {
 		case "enter":
-			m.ready = false
 			m.nextPage()
 			return m, nil
 		case "down":
@@ -79,6 +72,6 @@ func (m modelList) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return m, cmd
 }
 
-func (m modelList) View() string {
+func (m *modelList) View() string {
 	return m.list.View()
 }
