+++ internal/cli/views/items/file/picker/picker_test.go
@@ -3,7 +3,6 @@ package picker
 import (
 	"testing"
 
-	tea "github.com/charmbracelet/bubbletea"
 	"github.com/knz/catwalk"
 )
 
@@ -13,23 +12,10 @@ func (c *MockController) AddFile(string) error {
 	return nil
 }
 
-type PickerPageWrapper struct {
-	*PickerPage
-}
-
-func NewPickerPageWrapper() *PickerPageWrapper {
+func NewPickerPageWrapper() *modelPicker {
 	mp := newModelPicker(func() {}, &MockController{})
 	mp.testMode = true
-	pp := &PickerPage{mp, 0, 0}
-	return &PickerPageWrapper{pp}
-}
-
-func (m *PickerPageWrapper) Init() tea.Cmd {
-	return nil
-}
-
-func (m *PickerPageWrapper) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	return m.PickerPage.Update(m, msg)
+	return mp
 }
 
 func TestPicker(t *testing.T) {
