+++ internal/cli/views/items/pair/show/pairs
@@ -1,25 +1,6 @@
 run
 ----
 -- view:
-   Посмотреть пары логин/пароль   ␤
-                                  ␤
-No items.                         ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-                                  ␤
-    q quit • ? more               ␤
-                                  🛇
-
-run
-key enter
-----
--- view:
    Посмотреть пары логин/пароль        ␤
                                        ␤
   > 2020-01-01 03:00:00 +0300 MSK      ␤
